requirements.txt
# Google APIs
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Google Gemini AI
google-generativeai>=0.3.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0

